Add new modules & small refactor

+++ src/config/app.config.h
@@ -4,10 +4,10 @@ namespace SmartGardenIO::Config {
     constexpr unsigned long SerialBaudRate = 115200;
     constexpr unsigned long MainLoopIntervalMs = 2000;
 
-    constexpr uint8_t I2C_SerialData_PIN = 21;
-    constexpr uint8_t I2C_SerialClock_PIN = 22;
+    constexpr int I2C_SerialData_PIN = 21;
+    constexpr int I2C_SerialClock_PIN = 22;
 
-    constexpr uint8_t SoilMoisture_PIN = 34; 
+    constexpr int SoilMoisture_PIN = 34; 
     constexpr int SoilMoistureDryValue = 2550;
     constexpr int SoilMoistureWetValue = 1060;
